Citizen: rename EspooLogo to CityLogo
- Reflect current customizable implementation; Espoo-specific things are under lib-customizations
