+++ frontend/src/citizen-frontend/header/CityLogo.tsx
